ONOS-IDS-platform/netguard-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
pyotp==2.9.0
qrcode[pil]==8.0
httpx==0.28.1
python-multipart==0.0.17
cryptography==44.0.0
paramiko==3.5.0
reportlab==4.2.5